Split useSubmit
